Refactoring of auth content (#106)
There used to be just one authentication.md file under concepts, now there's an authentication root at authentication/authentication.md.

Will leave this branch to be merged with master too. There will be a small merge conflict but since this is a non-destructive change (if the redirect doesn't work the 404 error still occurs) it's easy to fix.