Updated curl sample for incoming webhooks
The curl command execution fails on Windows because it is a PowerShell alias for Invoke-WebCommand. Separated the example for macOS and Linux based systems and Windows. Windows has to call curl.exe to be 100% sure that curl is used. Furthermore, single quotes instead of double quotes were required to run the sample on Windows