Update add-authentication to detail AAD V2
This page should be brought in line with https://docs.microsoft.com/en-us/azure/bot-service/bot-builder-authentication?view=azure-bot-service-4.0&tabs=aadv2%2Ccsharp to allow AAD V2.  I added the steps to configure the v2 provider.  I also added a comment about defining scopes, since this is the problem I moved to V2 to solve.  Feel free to revise, but I thought I'd point out that the docs were out of sync.  I've tested this with the OAuthDialog, and it works as expected.