[Go Live][Aug 16] (#6691)
* Update auth-aad-sso.md

* Update conversation-messages.md

* Updates

* Updates

* Update conversation-messages.md

* Update apps-upload.md

* Update configuration-page.md

* Update apps-upload.md

* updates

* Update define-search-command.md

* Update channel-messages-with-rsc.md

Added link to API parameter which lists the replies of the specified message in a channel.

* Update channel-messages-with-rsc.md

updated image tags and alt-text. added preserve-view=true tag to link to specific version

* Update conversation-messages.md

* Minor edit

* Minor edit

...in content added via PR  https://github.com/MicrosoftDocs/msteams-docs/pull/5904

* Minor edits

* Update connectors-creating.md (#6689)

* [Doc Bug][video links] (#6690)

* video links

* video links fix

* Updated the articles

Edited and updated the alt+txt in the images. Also added lightbox image and hyperlinked the help articles in the manifest-schema module.

* Minor updates

Edited and updated the articles for acrolinx score

* Minor updates

Edited and updated the links in the manifest-schema article.

* Minor update

Deleted and added a screenshot of less pixel for better UX

* Minor update

Updated the image pixel and added to the article

* Minor update

Edited and updated the properties hyperlink

Co-authored-by: v-npaladugu <91744224+v-npaladugu@users.noreply.github.com>
Co-authored-by: v-habibaumme <91723607+v-habibaumme@users.noreply.github.com>
Co-authored-by: v-ypalikila <v-ypalikila@microsoft.com>
Co-authored-by: Ashish Gupta <guptaashish@microsoft.com>
Co-authored-by: Shalini Robinson <89063997+v-shalinirobinson@users.noreply.github.com>
Co-authored-by: Ajith kumar <v-ajiths@microsoft.com>