Update sbs-gs-notificationbot.yml (#7969)
* Update sbs-gs-notificationbot.yml

It's probably obvious to most people, but there's a ton of helpful information in the readme file of the project for what to do next and how to extend the bot.  NO ONE wants to just deploy this bot and be done. They need to know what's next. Just add something at the bottom of the page to let them know they can do a lot more with it and there is help in the readme file.

* Update sbs-gs-notificationbot.yml

* Update sbs-gs-notificationbot.yml

* Update sbs-gs-notificationbot.yml

* Update sbs-gs-notificationbot.yml

* Update sbs-gs-notificationbot.yml

---------

Co-authored-by: v-npaladugu <v-npaladugu@microsoft.com>