Added a note about Company communicator unavailability on mobile devices (#8809)
* Added a note about Company communicator unavailability on mobile devices

We have received an ICM recently where customer was trying to rin company communicator app on mobile but it was not working as expected. On further discussion we found that this is not yet supported on mobile. I think we could avoid a lot of to and fro by having this limitation documented.

* Update app-templates.md

* Update app-templates.md

* Update app-templates.md

* Update app-templates.md

---------

Co-authored-by: Surbhi-MSFT <75663981+surbhigupta12@users.noreply.github.com>
Co-authored-by: v-habibaumme <91723607+v-habibaumme@users.noreply.github.com>